complete chapter3

+++ Chapter3/task3/inc/queues.hpp
@@ -4,11 +4,11 @@
 class queues
 {
 private:
-	std::pair<uint, uint>	_clocks[3];
-	uint					_sizes[3];
+	std::pair<double, double>	_clocks[3];
+	uint						_sizes[3];
 public:
 	queues();
-	std::pair<uint, uint>	&operator[](uint index);
+	std::pair<double, double>	&operator[](uint index);
 	uint	getSize(uint index);
 	void	incSize(uint index);
 	void	decSize(uint index);
